钱包功能

+++ Creative_Market/src/main/java/com/creative/domain/buyType.java
@@ -11,5 +11,5 @@ public class buyType {
     private Integer buyId;
     private Integer commodityId;
     private String buyType;
-    private Integer buyMoney;
+    private Double buyMoney;
 }
